Move command subcommand tests to tests directory
Remove test/non-test conditional compilation in db.rs; keep only test code
