+++ libra/src/command/checkout.rs
@@ -30,7 +30,7 @@ pub async fn execute(args: CheckoutArgs) {
     }
 }
 
-async fn get_current_branch() -> Option<String> {
+pub async fn get_current_branch() -> Option<String> {
     let head = Head::current().await;
     match head {
         Head::Detached(commit_hash) => {
@@ -47,7 +47,7 @@ async fn show_current_branch() {
     }
 }
 
-async fn switch_branch(branch_name: &str) {
+pub async fn switch_branch(branch_name: &str) {
     let target_branch: Option<Branch> = Branch::find_branch(branch_name, None).await;
     let commit_id = target_branch.unwrap().commit;
     restore_to_commit(commit_id).await;
@@ -73,7 +73,7 @@ async fn get_remote(branch_name: &str) {
     pull::execute(pull::PullArgs::make(None, None)).await;
 }
 
-async fn check_branch(branch_name: &str) -> Option<bool> {
+pub async fn check_branch(branch_name: &str) -> Option<bool> {
     if get_current_branch().await == Some(branch_name.to_string()) {
         println!("Already on {branch_name}");
         return None;
@@ -119,92 +119,4 @@ async fn restore_to_commit(commit_id: SHA1) {
 
 /// Unit tests for the checkout module
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        command::{commit, init},
-        utils::test,
-    };
-    use colored::Colorize;
-    use serial_test::serial;
-    use tempfile::tempdir;
-
-    async fn test_check_branch() {
-        println!("\n\x1b[1mTest check_branch function.\x1b[0m");
-
-        // For non-existent branches, it should return None
-        assert_eq!(check_branch("non_existent_branch").await, None);
-        // For the current branch, it should return None
-        assert_eq!(
-            check_branch(&get_current_branch().await.unwrap_or("main".to_string())).await,
-            None
-        );
-        // For other existing branches, it should return Some(false)
-        assert_eq!(check_branch("new_branch_01").await, Some(false));
-    }
-
-    async fn test_switch_branch() {
-        println!("\n\x1b[1mTest switch_branch function.\x1b[0m");
-
-        let show_all_branches = async || {
-            // Use the list_branches function of the branch module to list all current local branches
-            branch::list_branches(false).await;
-            println!(
-                "Current branch is '{}'.",
-                get_current_branch()
-                    .await
-                    .unwrap_or("Get_current_branch_failed".to_string())
-                    .green()
-            );
-        };
-
-        // Switch to the new branch and back
-        show_all_branches().await;
-        switch_branch("new_branch_01").await;
-        show_all_branches().await;
-        switch_branch("new_branch_02").await;
-        show_all_branches().await;
-        switch_branch("main").await;
-        show_all_branches().await;
-    }
-
-    #[tokio::test]
-    #[serial]
-    /// Tests branch creation, switching and validation functionality in the checkout module.
-    /// Verifies proper branch management and HEAD reference updates when switching between branches.
-    async fn test_checkout_module_functions() {
-        println!("\n\x1b[1mTest checkout module functions.\x1b[0m");
-
-        let temp_path = tempdir().unwrap();
-        test::setup_clean_testing_env_in(temp_path.path());
-        let _guard = test::ChangeDirGuard::new(temp_path.path());
-
-        let init_args = init::InitArgs {
-            bare: false,
-            initial_branch: Some("main".to_string()),
-            repo_directory: temp_path.path().to_str().unwrap().to_string(),
-            quiet: false,
-        };
-
-        init::init(init_args)
-            .await
-            .expect("Error initializing repository");
-
-        // Initialize the main branch by creating an empty commit
-        let commit_args = commit::CommitArgs {
-            message: "An empty initial commit".to_string(),
-            allow_empty: true,
-            conventional: false,
-            amend: false,
-        };
-        commit::execute(commit_args).await;
-
-        // Create tow new branch
-        branch::create_branch(String::from("new_branch_01"), get_current_branch().await).await;
-        branch::create_branch(String::from("new_branch_02"), get_current_branch().await).await;
-
-        // Test the checkout module funsctions
-        test_check_branch().await;
-        test_switch_branch().await;
-    }
-}
+mod tests {}
