+++ libra/tests/command_test.rs
@@ -1,0 +1,2 @@
+//! Main integration test entry point
+mod command;
